An obscene amount of code refactoring
