+++ viewer/importers/__init__.py
@@ -1,0 +1,10 @@
+from viewer.importers.core import *
+from viewer.importers.dav import *
+from viewer.importers.git import *
+from viewer.importers.home_assistant import *
+from viewer.importers.location import *
+from viewer.importers.messages import *
+from viewer.importers.monica import *
+from viewer.importers.openscale import *
+from viewer.importers.photos import *
+
